refactor: moves DateRangePicker to a new file
